+++ src/lib.rs
@@ -17,6 +17,7 @@ mod hooks;
 mod row;
 mod rows;
 mod serialize;
+mod statement;
 mod storage;
 mod token;
 mod tosql;
@@ -45,6 +46,7 @@ pub use hooks::{Action, AuthContext, Authorization, CheckpointMode, TransactionO
 pub use row::{Row, RowIndex};
 pub use rows::{AndThenRows, MappedRows, Rows};
 pub use serialize::{deserialize as deserialize_database, serialize as serialize_database};
+pub use statement::Statement;
 pub use storage::{Database, Table};
 pub use token::{tokenize, Token, TokenError};
 pub use tosql::ToSql;